requirements.txt
astroid==2.9.3
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.4
coverage==6.3.2
distlib==0.3.4
django-environ==0.8.1
factory-boy==3.2.1
Faker==13.3.1
filelock==3.6.0
flake8==4.0.1
flake8-isort==4.1.1
greenlet==1.1.2
identify==2.4.11
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
pkg-resources==0.0.0
platformdirs==2.5.1
playwright==1.19.1
pluggy==1.0.0
pre-commit==2.17.0
py==1.11.0
pycodestyle==2.8.0
pyee==8.1.0
pyflakes==2.4.0
pylint==2.12.2
pylint-plugin-utils==0.7
pyparsing==3.0.7
pytest==7.0.1
pytest-base-url==1.4.2
pytest-cov==3.0.0
pytest-playwright==0.2.3
python-dateutil==2.8.2
python-slugify==6.1.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
testfixtures==6.18.5
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.2
typing-extensions==4.1.1
urllib3==1.26.8
virtualenv==20.13.3
websockets==10.1
wrapt==1.13.3
zipp==3.7.0
